Make ossl_time_now() a static inline function in include/internal/time.h
That makes the whole OSSL_TIME API usable from anywhere with no dependencies
other than publicly exposed types and functions.
