+++ crypto/build.info
@@ -99,7 +99,7 @@ $UTIL_COMMON=\
         threads_pthread.c threads_win.c threads_none.c initthread.c \
         context.c sparse_array.c asn1_dsa.c packet.c param_build.c \
         param_build_set.c der_writer.c threads_lib.c params_dup.c \
-        quic_vlint.c time.c
+        quic_vlint.c
 
 SOURCE[../libssl]=sparse_array.c
