+++ include/internal/time.h
@@ -12,6 +12,7 @@
 # pragma once
 
 # include <openssl/e_os2.h>     /* uint64_t */
+# include <openssl/err.h>
 # include "internal/e_os.h"     /* for struct timeval */
 # include "internal/safe_math.h"
 
@@ -59,9 +60,6 @@ uint64_t ossl_time2ticks(OSSL_TIME t)
     return t.t;
 }
 
-/* Get current time */
-OSSL_TIME ossl_time_now(void);
-
 /* The beginning and end of the time range */
 static ossl_unused ossl_inline
 OSSL_TIME ossl_time_zero(void)
@@ -75,6 +73,43 @@ OSSL_TIME ossl_time_infinite(void)
     return ossl_ticks2time(~(uint64_t)0);
 }
 
+/* Get current time */
+static ossl_unused ossl_inline
+OSSL_TIME ossl_time_now(void)
+{
+    OSSL_TIME r;
+
+#if defined(_WIN32)
+    SYSTEMTIME st;
+    union {
+        unsigned __int64 ul;
+        FILETIME ft;
+    } now;
+
+    GetSystemTime(&st);
+    SystemTimeToFileTime(&st, &now.ft);
+    /* re-bias to 1/1/1970 */
+# ifdef  __MINGW32__
+    now.ul -= 116444736000000000ULL;
+# else
+    now.ul -= 116444736000000000UI64;
+# endif
+    r.t = ((uint64_t)now.ul) * (OSSL_TIME_SECOND / 10000000);
+#else   /* defined(_WIN32) */
+    struct timeval t;
+
+    if (gettimeofday(&t, NULL) < 0) {
+        ERR_raise_data(ERR_LIB_SYS, get_last_sys_error(),
+                       "calling gettimeofday()");
+        return ossl_time_zero();
+    }
+    if (t.tv_sec <= 0)
+        r.t = t.tv_usec <= 0 ? 0 : t.tv_usec * OSSL_TIME_US;
+    else
+        r.t = ((uint64_t)t.tv_sec * 1000000 + t.tv_usec) * OSSL_TIME_US;
+#endif  /* defined(_WIN32) */
+    return r;
+}
 
 /* Convert time to timeval */
 static ossl_unused ossl_inline
